+++ Source/CDR.Register.API.Logger/RequestResponseLogger.cs
@@ -8,15 +8,11 @@
     public class RequestResponseLogger : IRequestResponseLogger, IDisposable
     {
         private readonly Logger _logger;
-        private readonly IConfiguration _configuration;
 
         public ILogger Log { get { return _logger; } }
 
         public RequestResponseLogger(IConfiguration configuration)
         {
-
-            _configuration = configuration;
-
             _logger = new LoggerConfiguration()
                 .ReadFrom.Configuration(configuration, sectionName: "SerilogRequestResponseLogger")
                 .Enrich.WithProperty("RequestMethod", "")